Cleanup visualization dir and reference images
