+++ test/python/visualization/test_circuit_text_drawer.py
@@ -12,6 +12,7 @@
 
 """ `_text_circuit_drawer` "draws" a circuit in "ascii art" """
 
+import os
 import unittest
 from codecs import encode
 from math import pi
@@ -30,6 +31,7 @@ from qiskit.extensions import UnitaryGate, HamiltonianGate
 from qiskit.circuit.library import (
     HGate,
     U2Gate,
+    U3Gate,
     XGate,
     CZGate,
     ZGate,
@@ -42,6 +44,7 @@ from qiskit.circuit.library import (
     CPhaseGate,
 )
 from qiskit.transpiler.passes import ApplyLayout
+from .visualization import path_to_diagram_reference, QiskitVisualizationTestCase
 
 
 class TestTextDrawerElement(QiskitTestCase):
@@ -1579,7 +1582,6 @@ class TestTextDrawerParams(QiskitTestCase):
         phi, lam = Parameter("φ"), Parameter("λ")
         circuit = QuantumCircuit(1)
         circuit.u2(phi, lam, 0)
-        print(circuit)
         self.assertEqual(circuit.draw(output="text").single_string(), expected)
 
 
@@ -3914,6 +3916,85 @@ class TestTextPhase(QiskitTestCase):
         circuit.global_phase = 4.21
 
         self.assertEqual(circuit.draw(output="text").single_string(), expected)
+
+
+class TestCircuitVisualizationImplementation(QiskitVisualizationTestCase):
+    """Tests utf8 and cp437 encoding."""
+
+    text_reference_utf8 = path_to_diagram_reference("circuit_text_ref_utf8.txt")
+    text_reference_cp437 = path_to_diagram_reference("circuit_text_ref_cp437.txt")
+
+    def sample_circuit(self):
+        """Generate a sample circuit that includes the most common elements of
+        quantum circuits.
+        """
+        qr = QuantumRegister(3, "q")
+        cr = ClassicalRegister(3, "c")
+        circuit = QuantumCircuit(qr, cr)
+        circuit.x(qr[0])
+        circuit.y(qr[0])
+        circuit.z(qr[0])
+        circuit.barrier(qr[0])
+        circuit.barrier(qr[1])
+        circuit.barrier(qr[2])
+        circuit.h(qr[0])
+        circuit.s(qr[0])
+        circuit.sdg(qr[0])
+        circuit.t(qr[0])
+        circuit.tdg(qr[0])
+        circuit.sx(qr[0])
+        circuit.sxdg(qr[0])
+        circuit.i(qr[0])
+        circuit.reset(qr[0])
+        circuit.rx(pi, qr[0])
+        circuit.ry(pi, qr[0])
+        circuit.rz(pi, qr[0])
+        circuit.append(U1Gate(pi), [qr[0]])
+        circuit.append(U2Gate(pi, pi), [qr[0]])
+        circuit.append(U3Gate(pi, pi, pi), [qr[0]])
+        circuit.swap(qr[0], qr[1])
+        circuit.cx(qr[0], qr[1])
+        circuit.cy(qr[0], qr[1])
+        circuit.cz(qr[0], qr[1])
+        circuit.ch(qr[0], qr[1])
+        circuit.append(CU1Gate(pi), [qr[0], qr[1]])
+        circuit.append(CU3Gate(pi, pi, pi), [qr[0], qr[1]])
+        circuit.crz(pi, qr[0], qr[1])
+        circuit.cry(pi, qr[0], qr[1])
+        circuit.crx(pi, qr[0], qr[1])
+        circuit.ccx(qr[0], qr[1], qr[2])
+        circuit.cswap(qr[0], qr[1], qr[2])
+        circuit.measure(qr, cr)
+
+        return circuit
+
+    def test_text_drawer_utf8(self):
+        """Test that text drawer handles utf8 encoding."""
+        filename = "current_textplot_utf8.txt"
+        qc = self.sample_circuit()
+        output = _text_circuit_drawer(
+            qc, filename=filename, fold=-1, initial_state=True, cregbundle=False, encoding="utf8"
+        )
+        try:
+            encode(str(output), encoding="utf8")
+        except UnicodeEncodeError:
+            self.fail("_text_circuit_drawer() should be utf8.")
+        self.assertFilesAreEqual(filename, self.text_reference_utf8, "utf8")
+        os.remove(filename)
+
+    def test_text_drawer_cp437(self):
+        """Test that text drawer handles cp437 encoding."""
+        filename = "current_textplot_cp437.txt"
+        qc = self.sample_circuit()
+        output = _text_circuit_drawer(
+            qc, filename=filename, fold=-1, initial_state=True, cregbundle=False, encoding="cp437"
+        )
+        try:
+            encode(str(output), encoding="cp437")
+        except UnicodeEncodeError:
+            self.fail("_text_circuit_drawer() should be cp437.")
+        self.assertFilesAreEqual(filename, self.text_reference_cp437, "cp437")
+        os.remove(filename)
 
 
 if __name__ == "__main__":
